scripts/news-bot/requirements.txt
feedparser==6.0.10
requests==2.32.3
python-dotenv==1.0.0
